# packages/schema/openapi.json
{
  "components": {
    "schemas": {
...
      "CreatePartnerIntegrationInput": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": ["https://example.com/schemas/CreatePartnerIntegrationInput.json"],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "allowed_scopes": {
            "items": {
              "type": "string"
            },
            "type": ["array", "null"]
          },
          "name": {
            "type": "string"
          }
        },
        "required": ["name", "allowed_scopes"],
        "type": "object"
      },
...
      "OpenAPIError": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": ["https://example.com/schemas/OpenAPIError.json"],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "error": {
            "type": "string"
          }
        },
        "required": ["error"],
        "type": "object"
      },
      "Organization": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": ["https://example.com/schemas/Organization.json"],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "external_account_id": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "partner_integration_id": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "scopes": {
            "items": {
              "type": "string"
            },
            "type": ["array", "null"]
          },
          "usage_level": {
            "type": "string"
          }
        },
        "required": ["id", "name", "usage_level", "created_at"],
        "type": "object"
      },
...
      "PartnerConnection": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "external_account_id": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "integration_id": {
            "type": "string"
          },
          "org_id": {
            "type": "string"
          },
          "revoked_at": {
            "format": "date-time",
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "integration_id",
          "external_account_id",
          "org_id",
          "status",
          "created_at"
        ],
        "type": "object"
      },
      "PartnerConnectionResponse": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": ["https://example.com/schemas/PartnerConnectionResponse.json"],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "access_token": {
            "type": "string"
          },
          "connection_id": {
            "type": "string"
          },
          "expires_in": {
            "format": "int64",
            "type": "integer"
          },
          "organization": {
            "$ref": "#/components/schemas/Organization"
          },
          "refresh_token": {
            "type": "string"
          },
          "scope": {
            "type": "string"
          },
          "token_type": {
            "type": "string"
          }
        },
        "required": [
          "connection_id",
          "organization",
          "access_token",
          "token_type",
          "expires_in",
          "refresh_token",
          "scope"
        ],
        "type": "object"
      },
      "PartnerIntegration": {
        "additionalProperties": false,
        "properties": {
          "allowed_scopes": {
            "items": {
              "type": "string"
            },
            "type": ["array", "null"]
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "disabled": {
            "type": "boolean"
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "owner_org_id": {
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": ["id", "owner_org_id", "name", "allowed_scopes", "created_at", "updated_at"],
        "type": "object"
      },
      "PartnerIntegrationCreated": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": ["https://example.com/schemas/PartnerIntegrationCreated.json"],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "allowed_scopes": {
            "items": {
              "type": "string"
            },
            "type": ["array", "null"]
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "disabled": {
            "type": "boolean"
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "owner_org_id": {
            "type": "string"
          },
          "secret": {
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "secret",
          "id",
          "owner_org_id",
          "name",
          "allowed_scopes",
          "created_at",
          "updated_at"
        ],
        "type": "object"
      },
      "PartnerRefreshRequest": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": ["https://example.com/schemas/PartnerRefreshRequest.json"],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "connection_id": {
            "type": "string"
          },
          "refresh_token": {
            "type": "string"
          }
        },
        "required": ["connection_id", "refresh_token"],
        "type": "object"
      },
...
      "ProvisionPartnerConnectionInput": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": ["https://example.com/schemas/ProvisionPartnerConnectionInput.json"],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "external_account_id": {
            "type": "string"
          },
          "organization_name": {
            "type": "string"
          },
          "requested_scopes": {
            "items": {
              "type": "string"
            },
            "type": ["array", "null"]
          }
        },
        "required": ["external_account_id", "organization_name", "requested_scopes"],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT or personal access token",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Nanoflare Control API",
    "version": "v1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/organizations/{orgID}/partner-integrations": {
      "get": {
        "operationId": "listPartnerIntegrations",
        "parameters": [
          {
            "in": "path",
            "name": "orgID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PartnerIntegration"
                  },
                  "type": ["array", "null"]
                }
              }
            },
            "description": "Success"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "listPartnerIntegrations",
        "tags": ["Partners"]
      },
      "post": {
        "operationId": "createPartnerIntegration",
        "parameters": [
          {
            "in": "path",
            "name": "orgID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePartnerIntegrationInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PartnerIntegrationCreated"
                }
              }
            },
            "description": "Success"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "createPartnerIntegration",
        "tags": ["Partners"]
      }
    },
    "/v1/organizations/{orgID}/partner-integrations/{integrationID}": {
      "delete": {
        "operationId": "disablePartnerIntegration",
        "parameters": [
          {
            "in": "path",
            "name": "orgID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "integrationID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No content"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "disablePartnerIntegration",
        "tags": ["Partners"]
      }
    },
    "/v1/organizations/{orgID}/partner-integrations/{integrationID}/client-secrets": {
      "post": {
        "operationId": "rotatePartnerIntegrationSecret",
        "parameters": [
          {
            "in": "path",
            "name": "orgID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "integrationID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PartnerIntegrationCreated"
                }
              }
            },
            "description": "Success"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "rotatePartnerIntegrationSecret",
        "tags": ["Partners"]
      }
    },
    "/v1/organizations/{orgID}/partner-integrations/{integrationID}/connections": {
      "get": {
        "operationId": "listPartnerConnections",
        "parameters": [
          {
            "in": "path",
            "name": "orgID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "integrationID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PartnerConnection"
                  },
                  "type": ["array", "null"]
                }
              }
            },
            "description": "Success"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "listPartnerConnections",
        "tags": ["Partners"]
      },
      "post": {
        "operationId": "provisionPartnerConnection",
        "parameters": [
          {
            "in": "path",
            "name": "orgID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "integrationID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProvisionPartnerConnectionInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PartnerConnectionResponse"
                }
              }
            },
            "description": "Success"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "provisionPartnerConnection",
        "tags": ["Partners"]
      }
    },
    "/v1/organizations/{orgID}/partner-integrations/{integrationID}/connections/{connectionID}": {
      "delete": {
        "operationId": "revokePartnerConnection",
        "parameters": [
          {
            "in": "path",
            "name": "orgID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "integrationID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "connectionID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No content"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "revokePartnerConnection",
        "tags": ["Partners"]
      }
    },
    "/v1/organizations/{orgID}/partner-integrations/{integrationID}/token": {
      "post": {
        "operationId": "refreshPartnerConnection",
        "parameters": [
          {
            "in": "path",
            "name": "orgID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "integrationID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PartnerRefreshRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PartnerConnectionResponse"
                }
              }
            },
            "description": "Success"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAPIError"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "refreshPartnerConnection",
        "tags": ["Partners"]
      }
    },
...
  }
}